Refactor sync locations DAG
Removed call to get a list of offices from the main body of the DAG
and moved it to a task.  NWDM and NWDP are not actual offices and
a seperate task is used to add them to the XCom.  RADAR json is not
the best formatted json and does need modifications.  Pandas is used
to do the mods with an additional method to take the Pandas json
normalization/drop duplicates and format the dictionary attributes
back to what's needed for the payload.