Refactor/a2w locations (#189)
* temporarily remove nbm dag to troubleshoot issue in develop

* add nbm back after deletion in develop

* Feature/airflow pytest (#179)

* WIP

* more WIP

* cleanup

* add GH workflow

* disable due to test failing (#180)

* WRF Columbia (#182)

* WIP: Requires updates to cumulus-api and geoproc

* Separate tasks to be same as BC DAGS

* Simplify slug naming

* Rename acquirable and retries is 1

* Matching acquirables to WRF BC

* Update UUID for WRF Columbia and BC

* WRF British Columbia (#181)

* WIP: Initial British Columbia DAG

* Simplify the slug naming

* Update acquirables name for British Columbia

* Rename acquirable and retiries is 1

* docstring updates

* Update UUID for WRF Columbia and BC

* WIP: refactor locations dag

* Refactor sync locations DAG

Removed call to get a list of offices from the main body of the DAG
and moved it to a task.  NWDM and NWDP are not actual offices and
a seperate task is used to add them to the XCom.  RADAR json is not
the best formatted json and does need modifications.  Pandas is used
to do the mods with an additional method to take the Pandas json
normalization/drop duplicates and format the dictionary attributes
back to what's needed for the payload.

* Added trigger rule to continue if a single office fails

* Remove unused imports

* Water API connection url change for local dev

Co-authored-by: Adam Scarberry <46494851+adamscarberry@users.noreply.github.com>