Sync CWMS Levels Refactor
Sync CWMS Levels DAG has both creating and updating in the same DAG.
A create levels list is created when not in Water and updated after
the create task finishes.

Full testing needed but looks like the DAG will work. . There seems
to be an issue with `v_locations`.  There are no entries for provider =
lrb.  Locations associated with LRB don't seem to be represented.